Move the alignment profile into its own module
This prevents circular dependencies between alignment, scorehandler,
and cli.
